% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_and_collapse.R
\name{find_top_variable_genes}
\alias{find_top_variable_genes}
\title{Finds top variable genes `TreeSE`}
\usage{
find_top_variable_genes(treeseobject, number)
}
\arguments{
\item{`TreeSE`}{S4 object of class TreeSummarizedExperiment}

\item{`number`}{Number of top genes to be calculated}
}
\value{
TreeSummarizedExperiment Object with added top_variable_gene information in metadata
}
\description{
Finds the n top variable genes within the genes present in Summarized Experiment object
}
\examples{


}
