% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_and_collapse.R
\name{reassign_and_collapse}
\alias{reassign_and_collapse}
\title{Reassigns Clusters in Seurat Object and Collapses Clustree}
\usage{
reassign_and_collapse(cluster_df, count_matrix)
}
\arguments{
\item{Dataframe}{containing cluster information at different resolutions}

\item{matrix}{Dense or sparse matrix containing the count matrix}
}
\value{
TreeSummarizedExperiment Object that can be visualized with metavizr
}
\description{
This module Reassigns Clusters in Seurat Object and Collapses Clustree.
}
\examples{

}
