% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_and_collapse.R
\name{simplified_treese}
\alias{simplified_treese}
\title{Creates the `TreeSE` object from provided hierarchy dataframe and count matrix}
\usage{
simplified_treese(cluster_df, count_matrix)
}
\arguments{
\item{Dataframe}{containing cluster information at different resolutions}

\item{matrix}{Dense or sparse matrix containing the count matrix}
}
\value{
TreeSummarizedExperiment Object that can be visualized with metavizr
}
\description{
This module returns the Tree Summarized Experiment if the dataframe is tree, throws error otherwise
}
\examples{

}
