% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_and_collapse.R
\name{visualizeSeurat}
\alias{visualizeSeurat}
\title{Wrapper for `Seurat`}
\usage{
visualizeSeurat(Seurat_object)
}
\arguments{
\item{Seurat_object}{S4 object of class Seurat containing cluster information at different resolutions}
}
\value{
TreeSummarizedExperiment Object that can be visualized with metavizr
}
\description{
This module Reassigns Clusters in Dataframe Object
}
\examples{

}
