% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_and_collapse.R
\name{visualizeSingleCellExperiment}
\alias{visualizeSingleCellExperiment}
\title{Wrapper for `SingleCellExperiment``}
\usage{
visualizeSingleCellExperiment(Sce_object)
}
\arguments{
\item{`SingleCellExperiment`}{S4 object of class Single Cell Experiment containing cluster information at different resolutions}
}
\value{
TreeSummarizedExperiment Object that can be visualized with metavizr
}
\description{
This module Reassigns Clusters in Dataframe Object
}
\examples{

}
